+++ R/slides.R
@@ -44,8 +44,7 @@ setMethod("[[", "SlideCatalog", function(x, i, ...) {
 
 #' @rdname catalog-extract
 #' @export
-setMethod(
-    "[[<-", c("SlideCatalog", "numeric", "missing", "CrunchSlide"),
+setMethod("[[<-", c("SlideCatalog", "numeric", "missing", "CrunchSlide"),
     function(x, i, j, value) {
         if (length(i) > 1) {
             # TODO, allow assignment of more than one slide
@@ -283,18 +282,19 @@ setMethod("[[", "AnalysisCatalog", function(x, i, ...) {
 })
 #' @rdname catalog-extract
 #' @export
-setMethod("[[<-", c("AnalysisCatalog", "numeric", "missing", "formula"), function(x, i, j, value) {
-    if (i > length(x)) {
-        halt("Index out of bounds, you can only assign a formula to an existing analysis.")
+setMethod("[[<-", c("AnalysisCatalog", "numeric", "missing", "formula"),
+    function(x, i, j, value) {
+        if (i > length(x)) {
+            halt("Index out of bounds, you can only assign a formula to an existing analysis.")
+        }
+        analysis <- x[[i]]
+        query(analysis) <- value
+        invisible(refresh(x))
     }
-    analysis <- x[[i]]
-    query(analysis) <- value
-    invisible(refresh(x))
-})
+)
 #' @rdname catalog-extract
 #' @export
-setMethod(
-    "[[<-", c("AnalysisCatalog", "numeric", "missing", "Analysis"),
+setMethod("[[<-", c("AnalysisCatalog", "numeric", "missing", "Analysis"),
     function(x, i, j, value) {
         if (length(i) > 1) {
             # TODO, recurse through i
@@ -316,8 +316,7 @@ setMethod(
 )
 #' @rdname catalog-extract
 #' @export
-setMethod(
-    "[[<-", c("AnalysisCatalog", "numeric", "missing", "list"),
+setMethod("[[<-", c("AnalysisCatalog", "numeric", "missing", "list"),
     function(x, i, j, value) {
         all_fmla <- vapply(value, function(x) inherits(x, "formula"), logical(1))
         if (any(!all_fmla)) {