+++ R/subvariables.R
@@ -26,11 +26,19 @@
 #' @name Subvariables
 #' @aliases Subvariables subvariables subvariables<-
 #' @seealso [`subvars-extract`] [`describe-catalog`] [`deleteSubvariable`] `vignette("array-variables", package="crunch")`
-NULL
+setGeneric("subvariables", function(x) standardGeneric("subvariables"))
+
+#' @rdname Subvariables
+setGeneric(
+    "subvariables<-",
+    function(x, value) standardGeneric("subvariables<-")
+)
 
 #' @rdname Subvariables
 #' @export
 setMethod("subvariables", "CategoricalArrayVariable", function(x) {
+    # TODO: it may be simpler and cleaner to just get the subvars from the
+    # entity rather than the subvariables catalog.
     tup <- tuple(x)
     catalog_url <- absoluteURL(tup$subvariables_catalog, base = tup@index_url)
     if (!is.null(tup$subreferences)) {
@@ -42,6 +50,10 @@ setMethod("subvariables", "CategoricalArrayVariable", function(x) {
     activeFilter(out) <- activeFilter(x)
     return(out)
 })
+
+#' @rdname Subvariables
+#' @export
+setMethod("subvariables", "CrunchVariable", function(x) NULL)
 
 subvariableURLs <- function(x) {
     return(absoluteURL(unlist(x$subvariables), base = x@index_url))
@@ -61,16 +73,14 @@ setMethod("subvariables", "VariableTuple", function(x) {
 
 #' @rdname Subvariables
 #' @export
-setMethod(
-    "subvariables<-", c("CategoricalArrayVariable", "ANY"),
+setMethod("subvariables<-", c("CategoricalArrayVariable", "ANY"),
     function(x, value) {
         halt("Can only assign an object of class Subvariables")
     }
 )
 #' @rdname Subvariables
 #' @export
-setMethod(
-    "subvariables<-", c("CategoricalArrayVariable", "Subvariables"),
+setMethod("subvariables<-", c("CategoricalArrayVariable", "Subvariables"),
     function(x, value) {
         old <- subvariableURLs(tuple(x))
         new <- urls(value)
@@ -133,9 +143,7 @@ setMethod("[", c("Subvariables", "character"), function(x, i, ...) {
 
 #' @rdname subvars-extract
 #' @export
-setMethod(
-    "[[<-",
-    c("Subvariables", "character", "missing", "CrunchVariable"),
+setMethod("[[<-", c("Subvariables", "character", "missing", "CrunchVariable"),
     function(x, i, value) {
         i <- match(i, names(x))
         if (is.na(i)) {
@@ -152,9 +160,7 @@ setMethod(
 )
 #' @rdname subvars-extract
 #' @export
-setMethod(
-    "[[<-",
-    c("Subvariables", "ANY", "missing", "CrunchVariable"),
+setMethod("[[<-", c("Subvariables", "ANY", "missing", "CrunchVariable"),
     function(x, i, value) {
         if (self(value) != urls(x)[i]) {
             halt("Cannot add or remove subvariables")
@@ -165,18 +171,14 @@ setMethod(
 )
 #' @rdname subvars-extract
 #' @export
-setMethod(
-    "[[<-",
-    c("Subvariables", "ANY", "missing", "NULL"),
+setMethod("[[<-", c("Subvariables", "ANY", "missing", "NULL"),
     function(x, i, value) {
         halt("Cannot add or remove subvariables")
     }
 )
 #' @rdname subvars-extract
 #' @export
-setMethod(
-    "[[<-",
-    c("Subvariables", "ANY", "missing", "ANY"),
+setMethod("[[<-", c("Subvariables", "ANY", "missing", "ANY"),
     function(x, i, value) {
         halt("Can only assign Variables into an object of class Subvariables")
     }
@@ -184,8 +186,7 @@ setMethod(
 
 #' @rdname subvars-extract
 #' @export
-setMethod(
-    "[<-", c("Subvariables", "character", "missing", "Subvariables"),
+setMethod("[<-", c("Subvariables", "character", "missing", "Subvariables"),
     function(x, i, value) {
         w <- match(i, names(x))
         if (any(is.na(w))) {
@@ -196,8 +197,7 @@ setMethod(
 )
 #' @rdname subvars-extract
 #' @export
-setMethod(
-    "[<-", c("Subvariables", "ANY", "missing", "Subvariables"),
+setMethod("[<-", c("Subvariables", "ANY", "missing", "Subvariables"),
     function(x, i, value) {
         inbound <- vapply(value, function(a) self(a), character(1))
         if (!all(inbound %in% urls(x)[i])) {
@@ -210,8 +210,7 @@ setMethod(
 )
 #' @rdname subvars-extract
 #' @export
-setMethod(
-    "[<-", c("Subvariables", "ANY", "missing", "ANY"),
+setMethod("[<-", c("Subvariables", "ANY", "missing", "ANY"),
     function(x, i, value) {
         halt("Can only assign Variables into an object of class Subvariables")
     }
@@ -272,9 +271,7 @@ setMethod("$", "CategoricalArrayVariable", function(x, name) x[[name]])
 
 #' @rdname subvars-extract
 #' @export
-setMethod(
-    "[[<-",
-    c("CategoricalArrayVariable", "ANY", "missing", "ANY"),
+setMethod("[[<-", c("CategoricalArrayVariable", "ANY", "missing", "ANY"),
     function(x, i, value) {
         subvariables(x)[[i]] <- value
         return(x)
@@ -282,9 +279,7 @@ setMethod(
 )
 #' @rdname subvars-extract
 #' @export
-setMethod(
-    "[[<-",
-    c("CategoricalArrayVariable", "character", "missing", "ANY"),
+setMethod("[[<-", c("CategoricalArrayVariable", "character", "missing", "ANY"),
     function(x, i, value) {
         i <- match(i, names(x))
         if (is.na(i)) {