+++ tests/testthat/test-dataset-entity.R
@@ -11,10 +11,21 @@ with_mock_crunch({
         expect_true(is.dataset(ds))
     })
 
+    test_that("CrunchDataset class (re-)init preserves object state", {
+        expect_identical(CrunchDataset(ds), ds)
+        expect_identical(CrunchDataset(ds[, "gender"]), ds[, "gender"])
+        expect_identical(CrunchDataset(ds[ds$gender == "Female",]),
+            ds[ds$gender == "Female",])
+        # Now confirm the same with subclasses
+        SubDataset <- setClass("SubDataset", contains="CrunchDataset")
+        expect_is(SubDataset(ds), "SubDataset")
+        expect_identical(names(SubDataset(ds[, "gender"])), "gender")
+    })
+
     test_that("Dataset attributes", {
         expect_identical(name(ds), "test ds")
         expect_identical(description(ds), "")
-        expect_identical(id(ds), "511a7c49778030653aab5963")
+        expect_identical(id(ds), "1")
         expect_null(notes(ds))
     })
 
@@ -310,18 +321,16 @@ with_mock_crunch({
         expect_identical(ds[url], ds["birthyr"])
     })
 
-    ## This is a start on a test that getting variables doesn't hit server.
-    ## It doesn't now, but if variable catalogs are lazily fetched, assert that
-    ## we're hitting cache
-    # with(temp.option(httpcache.log=""), {
-    #     dlog <- capture.output({
-    #         v1 <- ds$birthyr
-    #         d2 <- ds[names(ds)=="gender"]
-    #     })
-    # })
-    # print(dlog)
-    # logdf <- loadLogfile(textConnection(dlog))
-    # print(logdf)
+    test_that("Setting variable metadata on the dataset", {
+        # See other tests in test-variable-catalog.R; this tests the [<- methods
+        # on the dataset entity
+        expect_PATCH(names(variables(ds[1])) <- "year of birth",
+            'https://app.crunch.io/api/datasets/1/variables/', '{"element":"shoji:catalog","index":{',
+            '"https://app.crunch.io/api/datasets/1/variables/birthyr/":{',
+            '"name":"year of birth"}}}'
+        )
+        expect_no_request(names(variables(ds[1])) <- "Birth Year")
+    })
 
     test_that("Dataset extract error handling", {
         expect_error(ds[[999]], "subscript out of bounds")
@@ -454,6 +463,9 @@ with_mock_crunch({
         test_that("deleteDataset by URL", {
             expect_DELETE(deleteDataset(self(ds)), self(ds))
         })
+        test_that("deleteDataset by web URL", {
+            expect_DELETE(deleteDataset(APIToWebURL(ds)), self(ds))
+        })
         test_that("deleteDataset on Dataset object", {
             expect_DELETE(deleteDataset(ds), self(ds))
         })
@@ -476,6 +488,8 @@ with_mock_crunch({
                 "To delete, please identify the dataset uniquely by URL or path."
             )
         )
+        expect_error(deleteDataset("http://app.crunch.io/api/"),
+            "http://app.crunch.io/api/ is not a valid dataset URL")
     })
 
     test_that("Dashboard URL", {